feat: queryClient를 불러오는 getQueryClient함수 정의
Related to: #38
