+++ lib/getQueryClient.ts
@@ -1,0 +1,6 @@
+import { QueryClient } from '@tanstack/react-query';
+import { cache } from 'react';
+
+const getQueryClient = cache(() => new QueryClient());
+
+export default getQueryClient;
